Updates .gitignore iOS part compiles

+++ mopub-ios/MoPub/MoPub/MoPubTypeConversion.m
@@ -85,7 +85,7 @@
     }
     const char* utf8String = string.UTF8String;
     unsigned long length = strlen( utf8String );
-    return FRENewObjectFromUTF8( length + 1, (uint8_t*) utf8String, asObject );
+    return FRENewObjectFromUTF8( (uint32_t) length + 1, (uint8_t*) utf8String, asObject );
 }
 
 - (FREResult) FREGetInt:(int32_t)value asObject:(FREObject*)asObject
